+++ extensions/jira/src/components/FormSprintDropdown.tsx
@@ -1,6 +1,6 @@
 import { Form } from "@raycast/api";
 import { useCachedPromise } from "@raycast/utils";
-import { ForwardedRef, forwardRef, useState } from "react";
+import { ForwardedRef, forwardRef, useState, useMemo } from "react";
 
 import { getSprints } from "../api/sprints";
 
@@ -18,12 +18,24 @@ const FormSprintDropdown = forwardRef((props: FormSprintDropdownProps, ref: Forw
     [query]
   );
 
+  const formattedSprints = useMemo(
+    () =>
+      sprints?.map((sprint) => {
+        return {
+          ...sprint,
+          // Strip away the possible HTML tags and the numbers at the end
+          displayName: sprint.displayName.replace(/<[^>]*>/g, "").replace(/\s+\(\d+\)$/, ""),
+        };
+      }),
+    [sprints]
+  );
+
   return (
     <Form.Dropdown ref={ref} {...props} isLoading={isLoading} onSearchTextChange={setQuery} throttle>
       <Form.Dropdown.Item title="No Sprint" value="" />
 
-      {sprints?.map((sprint) => {
-        return <Form.Dropdown.Item key={sprint.id} title={`${sprint.name} (${sprint.state})`} value={`${sprint.id}`} />;
+      {formattedSprints?.map((sprint) => {
+        return <Form.Dropdown.Item key={sprint.value} title={sprint.displayName} value={sprint.value} />;
       })}
     </Form.Dropdown>
   );