+++ extensions/timers/src/startCustomTimer.tsx
@@ -1,6 +1,7 @@
 import { Action, ActionPanel, closeMainWindow, Form, getPreferenceValues, Toast } from "@raycast/api";
 import { useState } from "react";
-import { createCustomTimer, ensureCTFileExists, startTimer } from "./timerUtils";
+import { soundData } from "./soundData";
+import { checkForOverlyLoudAlert, createCustomTimer, ensureCTFileExists, startTimer } from "./timerUtils";
 import { CTInlineArgs, InputField, RayFormEvent, Values } from "./types";
 
 export default function CustomTimerView(props: { arguments: CTInlineArgs }) {
@@ -21,11 +22,13 @@ export default function CustomTimerView(props: { arguments: CTInlineArgs }) {
     } else if (isNaN(Number(values.seconds))) {
       setSecErr("Seconds must be a number!");
     } else {
+      if (!checkForOverlyLoudAlert()) return;
       closeMainWindow();
       const timerName = values.name ? values.name : "Untitled";
       const timeInSeconds = 3600 * Number(values.hours) + 60 * Number(values.minutes) + Number(values.seconds);
-      startTimer(timeInSeconds, timerName);
-      if (values.willBeSaved) createCustomTimer({ name: values.name, timeInSeconds: timeInSeconds });
+      startTimer(timeInSeconds, timerName, values.selectedSound);
+      if (values.willBeSaved)
+        createCustomTimer({ name: values.name, timeInSeconds: timeInSeconds, selectedSound: values.selectedSound });
     }
   };
 
@@ -123,6 +126,12 @@ export default function CustomTimerView(props: { arguments: CTInlineArgs }) {
           onBlur={item.validator}
         />
       ))}
+      <Form.Dropdown id="selectedSound" defaultValue="default" title="Sound">
+        <Form.Dropdown.Item value="default" title="Default" />
+        {soundData.map((item, index) => (
+          <Form.Dropdown.Item key={index} title={item.title} value={item.value} icon={item.icon} />
+        ))}
+      </Form.Dropdown>
       <Form.TextField id="name" title="Name" placeholder="Pour Tea" autoFocus={hasArgs} />
       <Form.Checkbox id="willBeSaved" label="Save as preset" />
     </Form>